Added ObsoleteCaseComponent
Not a super useful component in terms of Unity but it allows you to create ObsoleteAttributes which only get turned on if the Unity version is higher then the date they were marked. Unfortntely no errors for removed components will show up in your IDE and they will also not fail a build.