Refactored Weaver Assembly Resolver
The resolver was a little weak. It would just try to find any assemblies of Unity in the library folder. This caused a problem on Mac since the Unity assemblies are not copied there. Instead we take advantage of AppDoamin and whenever we need to resolve an assembly we look there. It would have to be loaded for it to exist in Unity.